Move fastly.wit into host_interface/wit and extract fastly interfaces
Fastly.wit has been broken up into individual compute-at-edge
interfaces, and the world definition has been extracted to its own file.
This change eases integration elsewhere by keeping the world definition
separate from the interface definition.
