+++ ci/clang-format.ignore
@@ -1,4 +1,4 @@
 runtime/js-compute-runtime/third_party/wizer.h
 runtime/js-compute-runtime/rust-url/rust-url.h
 runtime/js-compute-runtime/rust-encoding/rust-encoding.h
-runtime/js-compute-runtime/fastly-world/fastly_world.h
+runtime/js-compute-runtime/host_interface/component/fastly_world.h
