+++ runtime/js-compute-runtime/Makefile
@@ -151,9 +151,6 @@ $(BUILD):
 $(OBJ_DIR):
 	$(call cmd,mkdir,$@)
 
-$(OBJ_DIR)/fastly-world:
-	$(call cmd,mkdir,$@)
-
 $(OBJ_DIR)/impl:
 	$(call cmd,mkdir,$@)
 
@@ -167,6 +164,9 @@ $(OBJ_DIR)/core:
 	$(call cmd,mkdir,$@)
 
 $(OBJ_DIR)/host_interface:
+	$(call cmd,mkdir,$@)
+
+$(OBJ_DIR)/host_interface/component:
 	$(call cmd,mkdir,$@)
 
 shared:
@@ -277,7 +277,10 @@ regenerate-world:
 	@exit 1
 else
 regenerate-world:
-	$(WIT_BINDGEN) c fastly.wit --no-helpers --out-dir fastly-world
+	$(WIT_BINDGEN) c --no-helpers \
+	  --out-dir host_interface/component \
+	  --world fastly-world \
+	  host_interface/wit
 endif
 
 
@@ -295,7 +298,8 @@ $(foreach source,$(FSM_CPP),$(eval $(call compile_cxx,$(source))))
 
 # Compute runtime build ########################################################
 
-$(eval $(call compile_cxx,$(FSM_SRC)/fastly-world/fastly_world_adapter.cpp))
+$(eval $(call compile_cxx,$(FSM_SRC)/host_interface/component/fastly_world_adapter.cpp))
+$(eval $(call compile_cxx,$(FSM_SRC)/impl/main.cpp))
 
 # This version of the runtime uses the world adapter in place of the world
 # generated by wit-bindgen, adapting to the currently available host calls on
@@ -306,17 +310,16 @@ $(eval $(call compile_cxx,$(FSM_SRC)/fastly-world/fastly_world_adapter.cpp))
 # that script for more information about why we do this.
 $(OBJ_DIR)/js-compute-runtime.wasm: $(FSM_OBJ) $(SM_OBJ) $(RUST_URL_LIB) $(RUST_ENCODING_LIB)
 $(OBJ_DIR)/js-compute-runtime.wasm: $(OBJ_DIR)/impl/main.o
-$(OBJ_DIR)/js-compute-runtime.wasm: $(OBJ_DIR)/fastly-world/fastly_world_adapter.o
+$(OBJ_DIR)/js-compute-runtime.wasm: $(OBJ_DIR)/host_interface/component/fastly_world_adapter.o
 	$(call cmd_format,WASI_LD,$@) PATH="$(FSM_SRC)/scripts:$$PATH" \
 	$(WASI_CXX) $(LD_FLAGS) $(OPENSSL_LIBS) -o $@ $^
 	$(call cmd_format,WASM_STRIP,$@) $(call WASM_STRIP,$@)
 	$(call cmd_format,WASM_METADATA,$@) $(call WASM_METADATA,$@)
 
-$(eval $(call compile_cxx,$(FSM_SRC)/impl/main.cpp))
-
 # Compute runtime component build ##############################################
 
-$(eval $(call compile_c,$(FSM_SRC)/fastly-world/fastly_world.c))
+$(eval $(call compile_c,$(FSM_SRC)/host_interface/component/fastly_world.c))
+$(eval $(call compile_cxx,$(FSM_SRC)/impl/main_component.cpp))
 
 # This version of the runtime uses the code generated by wit-bindgen, to
 # ultimately be run as a component on c@e.
@@ -326,14 +329,12 @@ $(eval $(call compile_c,$(FSM_SRC)/fastly-world/fastly_world.c))
 # that script for more information about why we do this.
 $(OBJ_DIR)/js-compute-runtime-component.wasm: $(FSM_OBJ) $(SM_OBJ) $(RUST_URL_LIB) $(RUST_ENCODING_LIB)
 $(OBJ_DIR)/js-compute-runtime-component.wasm: $(OBJ_DIR)/impl/main_component.o
-$(OBJ_DIR)/js-compute-runtime-component.wasm: $(FSM_SRC)/fastly-world/fastly_world_component_type.o
-$(OBJ_DIR)/js-compute-runtime-component.wasm: $(OBJ_DIR)/fastly-world/fastly_world.o
+$(OBJ_DIR)/js-compute-runtime-component.wasm: $(FSM_SRC)/host_interface/component/fastly_world_component_type.o
+$(OBJ_DIR)/js-compute-runtime-component.wasm: $(OBJ_DIR)/host_interface/component/fastly_world.o
 	$(call cmd_format,WASI_LD,$@) PATH="$(FSM_SRC)/scripts:$$PATH" \
 	$(WASI_CXX) $(LD_FLAGS) $(OPENSSL_LIBS) -o $@ $^
 	$(call cmd_format,WASM_STRIP,$@) $(call WASM_STRIP,$@)
 	$(call cmd_format,WASM_METADATA,$@) $(call WASM_METADATA,$@)
-
-$(eval $(call compile_cxx,$(FSM_SRC)/impl/main_component.cpp))
 
 # Shared builtins build ########################################################
