+++ runtime/js-compute-runtime/host_interface/component/fastly_world.h
