+++ runtime/js-compute-runtime/host_interface/component/fastly_world_adapter.cpp
